+++ pkg/release/bumpall.go
@@ -554,6 +554,31 @@ func findNewestApp(name string, getUpstreamVersion bool) (appVersion, error) {
 		if err != nil {
 			return appVersion{}, microerror.Mask(err)
 		}
+	case "azuredisk-csi-driver":
+		version, err = getLatestGithubRelease("giantswarm", "azuredisk-csi-driver-app")
+		if err != nil {
+			return appVersion{}, microerror.Mask(err)
+		}
+	case "azurefile-csi-driver":
+		version, err = getLatestGithubRelease("giantswarm", "azurefile-csi-driver-app")
+		if err != nil {
+			return appVersion{}, microerror.Mask(err)
+		}
+	case "azure-cloud-node-manager":
+		version, err = getLatestGithubRelease("giantswarm", "azure-cloud-node-manager-app")
+		if err != nil {
+			return appVersion{}, microerror.Mask(err)
+		}
+	case "azure-cloud-controller-manager":
+		version, err = getLatestGithubRelease("giantswarm", "azure-cloud-controller-manager-app")
+		if err != nil {
+			return appVersion{}, microerror.Mask(err)
+		}
+	case "os-tooling":
+		version, err = getLatestGithubRelease("giantswarm", "capi-image-builder")
+		if err != nil {
+			return appVersion{}, microerror.Mask(err)
+		}
 
 	default:
 		version, err = getLatestGithubRelease("giantswarm", name)